requirements.txt
audio-recorder-streamlit
google-cloud-texttospeech
langchain-ollama
langchain-community
pypdf